Use new trait TokenSink for rule actions
The TokenSink trait is introduced to handle rule actions in a
more flexible manner. There is no longer any particular requirement
on rule signatures; instead the TokenSink is called by the
interpreter with the rule number of the current match. It is up
to the sink to decide what to do with the rule number (and other
token data passed in from the interpreter).