Abandon use of TokenSink in regex API
The regex engine now just returns a MatchRecord when requested.
The caller then worries about how to respond.

Most changes were motivated by a desire to get the regex patterns
into the same module as the tokenizer actions, so that a particular
tokenizer (e.g., for general English) is all in one place.

The borrow checker gave lots of problems, and that was a major
motivator in splitting the matching logic off from the action
logic. Basically, the problem is that the outer shell deals in
AnnotatedDocuments, while the inner core deals in matching a
string and updating a "TokenSink". So that means we have to
split up the AD into its text and its markup objects, and pass
them separately, basically by assigning temporary mutable
ownership of the markup objects to the TokenSink. But then we
cannot simultaneously get a reference to the text!

Better to treat the regex matcher as an encapsulated service,
that tells its caller about the tokens that it finds, but
takes no action on its own.